requirements.txt
Django==3.1.7
django-annoying==0.10.6
django-filter==2.4.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-jsonapi==4.0.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
Jinja2==2.11.3
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==2.0.1
social-auth-app-django==4.0.0
social-auth-core==4.0.3
uWSGI==2.0.19.1
celery==5.0.5
django_celery_beat==2.2.0
django-celery-results==2.0.1